consolodating astro and tsx versions
