+++ src/apps/pages/sections/FreeText.astro
@@ -1,6 +1,6 @@
 ---
 import { TinaMarkdown } from 'tinacms/dist/rich-text';
-import Spacer from './mdx/Spacer';
+import Spacer from '@components/Spacer';
 
 const { content } = Astro.props;
 const components = {
